requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
fastapi==0.135.0
filelock==3.24.3
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jiter==0.13.0
keyring==25.6.0
llvmlite==0.46.0
lxml==6.0.2
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
openai==2.24.0
openai-whisper==20250625
pillow==11.0.0
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.1.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.22.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pymem==1.14.0
pyparsing==3.3.2
pypng==0.20220715.0
PyQRCode==1.2.1
python-dotenv==1.2.1
python-multipart==0.0.22
pywin32==310
pywin32-ctypes==0.2.3
qrcode==8.2
regex==2026.2.28
requests==2.31.0
rsa==4.9.1
setuptools==82.0.0
sniffio==1.3.1
starlette==0.52.1
sympy==1.14.0
tiktoken==0.12.0
torch==2.10.0
tqdm==4.66.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uv==0.9.15
uvicorn==0.41.0
